tests/wpt/web-platform-tests/preload/link_header_preload.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that Link headers preload resources');
</script>
<body>
<script src="resources/dummy.js?pipe=trickle(d5)"></script>
<script>
    window.addEventListener("load", t.step_func(function() {
        var entries = performance.getEntriesByType("resource");
        assert_equals(entries.length, 6);
        t.done();
    }));
</script>
</body>
